requirements.txt
django
django-crispy-forms
social-auth-app-django